Refactor: 프로젝트 구조 변경
- 패키지 구분
- 불필요한 Fragment 삭제
