+++ app/src/main/java/com/survivalcoding/stopwatch/presentation/fragment/OtherFragment.kt
@@ -1,0 +1,5 @@
+package com.survivalcoding.stopwatch.presentation.fragment
+
+import androidx.fragment.app.Fragment
+
+class OtherFragment : Fragment()
