--- app/src/main/java/com/survivalcoding/stopwatch/ui/fragment/AlarmFragment.kt
@@ -1,6 +1,0 @@
-package com.survivalcoding.stopwatch.ui.fragment
-
-import androidx.fragment.app.Fragment
-
-class AlarmFragment : Fragment() {
-}
