--- app/src/main/java/com/survivalcoding/stopwatch/ui/fragment/ClockFragment.kt
@@ -1,6 +1,0 @@
-package com.survivalcoding.stopwatch.ui.fragment
-
-import androidx.fragment.app.Fragment
-
-class ClockFragment : Fragment() {
-}
