+++ app/src/main/res/navigation/navigation_graph.xml
@@ -7,31 +7,31 @@
 
     <fragment
         android:id="@+id/alarm"
-        android:name="com.survivalcoding.stopwatch.ui.fragment.AlarmFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.OtherFragment"
         android:label="@string/menu_alarm"
-        tools:layout="@layout/fragment_alarm" />
+        tools:layout="@layout/fragment_other" />
 
     <fragment
         android:id="@+id/clock"
-        android:name="com.survivalcoding.stopwatch.ui.fragment.ClockFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.OtherFragment"
         android:label="@string/menu_clock"
-        tools:layout="@layout/fragment_clock" />
+        tools:layout="@layout/fragment_other" />
 
     <fragment
         android:id="@+id/timer"
-        android:name="com.survivalcoding.stopwatch.ui.fragment.TimerFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.OtherFragment"
         android:label="@string/menu_timer"
-        tools:layout="@layout/fragment_timer" />
+        tools:layout="@layout/fragment_other" />
 
     <fragment
         android:id="@+id/stopwatch"
-        android:name="com.survivalcoding.stopwatch.ui.fragment.StopWatchFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.StopWatchFragment"
         android:label="@string/menu_stopwatch"
         tools:layout="@layout/fragment_stop_watch" />
 
     <fragment
         android:id="@+id/bedtime"
-        android:name="com.survivalcoding.stopwatch.ui.fragment.BedTimeFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.OtherFragment"
         android:label="@string/menu_bedtime"
-        tools:layout="@layout/fragment_bed_time" />
+        tools:layout="@layout/fragment_other" />
 </navigation>
